First refactor attempt
