--- .travis.yml
@@ -1,16 +1,0 @@
-services:
-  - docker
-
-install:
-  #- git clone --quiet --depth 1 https://github.com/mikaelarguedas/ros2ci.git .ros2ci
-  - git clone --quiet --depth 1 --branch mocap4ros_vicon https://github.com/IntelligentRoboticsLabs/ros2ci.git .ros2ci
-  # optional: add custom repos or rosinstall file to the workspace for all jobs
-  # these will be compiled in an underlay workspace
-  # - cp my_additional_repos.repos .ros2ci/additional_repos.repos
-
-matrix:
-  include:
-    - env: JOB_TYPE=eloquent
-      stage: build
-      name: "Build"
-      script: .ros2ci/travis.bash $JOB_TYPE
