+++ README.md
@@ -4,122 +4,6 @@
 
 This project provides support for ROS2 integration with Vicon cameras (MOCAP systems based on vision) and Technaid TechMCS IMUs (MOCAP systems based on motion sensors).
 
-
-# Install 
-
-Before installing the project you will have to install Vicon DataStream SDK.
-
-## Vicon DataStream SDK
-
-Get the official binaries released in the official download page [here](https://www.vicon.com/software/datastream-sdk/?section=downloads). 
-Last version used in this project was 1.11.0.
-
-You can check the documentation [here](https://docs.vicon.com/spaces/viewspace.action?key=DSSDK19).
-
-### Installing on Linux (version 1.11.0)
-
-After cloning the repository (see next section), copy the SDK zip file (`ViconDataStreamSDK_1.11.0_128037.zip`) to the folder `vicon2_driver` and use `vicon_sdk_install.sh` script. You will need p7zip installed on your system:
-```bash
-sudo aptitude install p7zip-full
-```
-
-### Installing on Windows
-
-Uncompress the official binaries, uncompress the corresponding Windows binaries (typically Win64) and run the MSI installer `ViconDataStreamSDK_1.11.0.128037h__x64.msi`.
-
-SDK files are placed to `C:\Program Files\Vicon\DataStream SDK\Win64\CPP` by default, so update the `PATH` environment variable with this SDK path:
-
-`set PATH=%PATH%;C:\Program Files\Vicon\DataStream SDK\Win64\CPP`
-
-
-
-## ROS2 Nodes
-
-### ROS2 Humble install under Linux
-Once the SDK zip file is downloaded, we will install this repo and [mocap](https://github.com/MOCAP4ROS2-Project/mocap.git) repository, which is also a dependency. 
-
-First, we clone the repositories:
-```bash
-cd ~/workspace
-mkdir -p mocap/src
-cd mocap/src
-git clone https://github.com/MOCAP4ROS2-Project/mocap_msgs.git
-git clone https://github.com/MOCAP4ROS2-Project/mocap.git
-git clone https://github.com/MOCAP4ROS2-Project/mocap4ros2_vicon.git
-```
-
-Next, we copy the SDK zip file to the vicon2_driver folder and run our installer:
-```bash
-cd ~/workspace/mocap/src/mocap4ros2_vicon/vicon2_driver
-cp ~/Downloads/ViconDataStreamSDK_1.11.0_128037.zip .
-./vicon_sdk_install.sh
-```
-
-Now, we install any missing dependency:
-```bash
-cd ~/workspace/mocap/src
-rosdep install --from-paths . --ignore-src --rosdistro humble -r -y 
-```
-
-And compile:
-```bash
-cd ~/workspace/mocap/
-colcon build --symlink-install
-```
-
-Now you can launch MOCAP4ROS2 nodes. Remember to "source" this Vicon workspace before with one of these commands:
-```bash
-# Linux
-source ~/workspace/mocap/install.setup.bash
-
-# Windows
-call local_setup.bat
-```
-
-
-# Running MOCAP4ROS2
-
-## Set-up
-- Connect your ROS2 computer/laptop to the same network as Vicon system is connected.
-- Run Nexus (Vicon software) and calibrate cameras (if required).
-- Edit some parameters on configuration file (`mocap4ros2_drivers/vicon2_driver/config/vicon2_driver_params.yaml`) to match your setup: e.g. `host_name` parameter.
-
-## Starting  Vicon-ROS2 driver
-- `vicon2_driver` is a lifecycle node. It can be started using the provided launcher:
-```bash
-ros2 launch vicon2_driver vicon2.launch.py
-```
-Once ready, we can activate it using different states. See next section for details about it. 
-
-- `vicon2_driver` node works at the frequency established by the Vicon System.
-- It will provide two publishers:
-    - One with the markers at  `[tracked_frame_suffix_]/markers`
-    - One TransformBroadcaster sending TFs to `/tf`
-
-## Lifecyclenode state management
-- `vicon2_driver` is a lifecycle node with different states, to know the different states you can run the next command in a terminal: 
-```bash
-ros2 lifecycle list vicon2_driver_node
-- cleanup [2]
-	Start: inactive
-	Goal: cleaningup
-- activate [3]
-	Start: inactive
-	Goal: activating
-- shutdown [6]
-	Start: inactive
-	Goal: shuttingdown
-```
-- Once you have launched the `vicon2_driver` you can activate it by running in other terminal:
-```bash
-ros2 lifecycle set vicon2_driver_node activate
-```
-
-- If you want to stop the vicon2_driver you have to run in other terminal:
-```bash
-ros2 lifecycle set vicon2_driver_node shutdown
-```
-
 # About MOCAP4ROS2
