--- vicon2_driver/package.xml
@@ -1,42 +1,0 @@
-<package format="3">
-  <name>vicon2_driver</name>
-  <version>0.0.1</version>
-  <description>
-     This is a ROS2 driver providing data from VICON motion capture systems.
-  </description>
-  <author>David Vargas</author>
-  <maintainer email="david.vargas@urjc.es">David Vargas</maintainer>
-
-  <license>BSD</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <build_depend>rosidl_default_generators</build_depend>
-  <build_depend>rclcpp</build_depend>
-  <build_depend>rclcpp_lifecycle</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>tf2</build_depend>
-  <build_depend>tf2_msgs</build_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>mocap_msgs</build_depend>
-  <build_depend>mocap_control</build_depend>
-
-  <exec_depend>geometry_msgs</exec_depend>
-  <exec_depend>rclcpp</exec_depend>
-  <exec_depend>rclcpp_lifecycle</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-  <exec_depend>tf2</exec_depend>
-  <exec_depend>tf2_msgs</exec_depend>
-  <exec_depend>geometry_msgs</exec_depend>
-  <exec_depend>mocap_msgs</exec_depend>
-  <exec_depend>mocap_control</exec_depend>
-
-  <test_depend>ament_cmake_gmock</test_depend>
-  <test_depend>ament_cmake_gtest</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  <test_depend>ament_index_cpp</test_depend>
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
